requirements.txt
ccxt==3.0.78
fastapi==0.99.1
numpy==1.23.3
pandas==2.0.1
plotly==5.14.1
psycopg2_binary==2.9.6
pyTelegramBotAPI==4.12.0
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0
scipy==1.9.1
